Move runtime code to a new zserio crate
