+++ Cargo.toml
@@ -1,11 +1,21 @@
 [workspace]
 
 members = [
+    "crates/zserio",
     ".",
     "tests/reference-module-lib",
     "tests/round-trip-tests",
     "tests/compare-ref-impl-tests/rust",
 ]
+
+[workspace.package]
+version = "0.1.0"
+edition = "2021"
+authors = ["Clemens Zangl <clemens.zangl@gmail.com>"]
+documentation = "https://docs.rs/rust-zserio"
+homepage = "https://github.com/Danaozhong/rust-zserio"
+repository = "https://github.com/Danaozhong/rust-zserio"
+license = "BSD-3-Clause"
 
 [package]
 name = "rust-zserio"
@@ -29,24 +39,15 @@ debug = true
 
 [dependencies]
 antlr-rust = "0.3.0-beta"
-bitmask-enum = "2.2.4"
-bitreader = "0.3.8"
-bitstream-io = "1.6.0"
 clap = { version = "4.2.7", features = ["derive"] }
 codegen = "0.2.0"
-duplicate = "1.0.0"
-half = "2.3.1"
 itertools = "0.11.0"
 log = "0.4.22"
-num = "0.4.0"
 prettyplease = "0.2.20"
-proc-macro2 = "1.0.66"
-rstest = "0.16.0"
-rust-bitwriter = "0.0.1"
 stringcase = "0.3.0"
 syn = "2.0.74"
-thiserror = "1.0.61"
 walkdir = "2.3.2"
+zserio = { version = "0.1.0", path = "crates/zserio" }
 
 [dependencies.simple_logger]
 version = "5.0.0"
