+++ crates/zserio/src/doctest.rs
