+++ crates/zserio/src/error.rs
