+++ crates/zserio/src/ztype/alignment.rs
