+++ crates/zserio/src/ztype/array_traits/boolean_array_trait.rs
