+++ crates/zserio/src/ztype/array_traits/delta_context.rs
