+++ crates/zserio/src/ztype/array_traits/object_array_trait.rs
