+++ crates/zserio/src/ztype/array_traits/varint16_array_trait.rs
