+++ crates/zserio/src/ztype/array_traits/varint32_array_trait.rs
