+++ crates/zserio/src/ztype/array_traits/varint_array_trait.rs
