+++ crates/zserio/src/ztype/array_traits/varuint32_array_trait.rs
