+++ crates/zserio/src/ztype/array_traits/varuint_array_trait.rs
