+++ crates/zserio/src/ztype/bit_alignment.rs
