+++ crates/zserio/src/ztype/float_encode.rs
