+++ crates/zserio/src/ztype/string_decode.rs
