+++ crates/zserio/src/ztype/string_encode.rs
