+++ crates/zserio/src/ztype/traits.rs
