+++ crates/zserio/src/ztype/varint_bitsize.rs
