+++ crates/zserio/src/ztype/varint_encode.rs
