+++ crates/zserio/src/ztype/varuint_bitsize.rs
