+++ crates/zserio/src/ztype/varuint_encode.rs
