+++ crates/zserio/src/ztype/writer.rs
