--- src/ztype.rs
@@ -1,95 +1,0 @@
-pub use self::string_encode::bitsize_string;
-pub use self::string_encode::write_string;
-
-pub use self::string_decode::read_string;
-
-pub use self::bits_encode::write_int16;
-pub use self::bits_encode::write_int32;
-pub use self::bits_encode::write_int64;
-pub use self::bits_encode::write_int8;
-
-pub use self::bits_encode::write_uint16;
-pub use self::bits_encode::write_uint32;
-pub use self::bits_encode::write_uint64;
-pub use self::bits_encode::write_uint8;
-
-pub use self::bits_encode::write_signed_bits;
-pub use self::bits_encode::write_unsigned_bits;
-
-pub use self::bits_decode::read_int16;
-pub use self::bits_decode::read_int32;
-pub use self::bits_decode::read_int64;
-pub use self::bits_decode::read_int8;
-
-pub use self::bits_decode::read_uint16;
-pub use self::bits_decode::read_uint32;
-pub use self::bits_decode::read_uint64;
-pub use self::bits_decode::read_uint8;
-
-pub use self::bits_decode::read_signed_bits;
-pub use self::bits_decode::read_unsigned_bits;
-
-pub use self::varuint_bitsize::{
-    unsigned_bitsize, varsize_bitsize, varuint16_bitsize, varuint32_bitsize, varuint64_bitsize,
-    varuint_bitsize,
-};
-
-pub use self::varuint_encode::{
-    write_varsize, write_varuint, write_varuint16, write_varuint32, write_varuint64,
-};
-
-pub use self::varuint_decode::{
-    read_varsize, read_varuint, read_varuint16, read_varuint32, read_varuint64,
-};
-
-pub use self::varint_bitsize::{
-    signed_bitsize, varint16_bitsize, varint32_bitsize, varint64_bitsize, varint_bitsize,
-};
-
-pub use self::varint_encode::{write_varint, write_varint16, write_varint32, write_varint64};
-
-pub use self::varint_decode::{read_varint, read_varint16, read_varint32, read_varint64};
-
-pub use self::float_encode::{write_float16, write_float32, write_float64};
-
-pub use self::float_decode::{read_float16, read_float32, read_float64};
-
-pub use self::array::Array;
-
-pub use self::bit_alignment::align_to;
-pub use self::traits::ZserioPackableObject;
-
-pub use self::bytes_type::{bitsize_of_bytes, read_bytes_type, write_bytes_type, BytesType};
-pub use self::extern_type::{bitsize_of_extern, read_extern_type, write_extern_type, ExternType};
-
-pub use self::numbits::{bit_length, numbits};
-
-pub use self::alignment::{align_bitsize, align_reader, align_writer};
-pub mod bits_decode;
-pub mod bits_encode;
-pub mod string_decode;
-pub mod string_encode;
-
-pub mod varint_bitsize;
-pub mod varint_decode;
-pub mod varint_encode;
-
-pub mod varuint_bitsize;
-pub mod varuint_decode;
-pub mod varuint_encode;
-
-// all the modules required for arrays
-pub mod array;
-pub mod array_traits;
-pub mod bit_alignment;
-
-pub mod alignment;
-pub mod bytes_type;
-pub mod extern_type;
-pub mod float_decode;
-pub mod float_encode;
-pub mod limits;
-pub mod numbits;
-pub mod reader;
-pub mod traits;
-pub mod writer;
